Увеличена точность кластеризации

+++ WebUI/BusinessLogic.Dtos/Cluster.cs
@@ -6,15 +6,15 @@ namespace BusinessLogic.Dtos
 {
     public class Cluster
     {
-        public Dictionary<ChannelEnum, float> CenterCluster { get; set; }
+        public Dictionary<ChannelEnum, double> CenterCluster { get; set; }
         public IEnumerable<ClusterPoint> Points { get; set; }
         public bool IsJoined { get; set; }
-        public float Ndvi { get; set; }
+        public double Ndvi { get; set; }
         public Color ClusterColor { get; set; }
 
         public Cluster()
         {
-            CenterCluster = new Dictionary<ChannelEnum, float>();
+            CenterCluster = new Dictionary<ChannelEnum, double>();
             Points = new List<ClusterPoint>();
             IsJoined = false;
         }
