+++ WebUI/BusinessLogic.Managers/ClassificationManager.cs
@@ -25,12 +25,12 @@ namespace BusinessLogic.Managers
         /// <summary>
         /// Параметр, характеризующий среднеквадратическое отклонение
         /// </summary>
-        private float _tettaS;
+        private double _tettaS;
 
         /// <summary>
         /// Параметр, характеризующий компактность
         /// </summary>
-        private float _tettaC;
+        private double _tettaC;
 
         /// <summary>
         /// Максимальное количество пар центров кластеров, которые можно объединить
@@ -50,33 +50,33 @@ namespace BusinessLogic.Managers
         /// <summary>
         /// Среднее расстояние между объектами входящих в кластер
         /// </summary>
-        private readonly IList<float> _dj = new List<float>();
+        private readonly IList<double> _dj = new List<double>();
 
         /// <summary>
         /// Обобщенное среднее расстояние между объектами, находящимися в отдельных кластерах, и соответствующими
         /// центрами кластеров
         /// </summary>
-        private float _d;
+        private double _d;
 
         /// <summary>
         /// Вектор среднеквадратичного отколнения
         /// </summary>
-        private readonly IList<Dictionary<ChannelEnum, float>> _sigmaj = new List<Dictionary<ChannelEnum, float>>();
+        private readonly IList<Dictionary<ChannelEnum, double>> _sigmaj = new List<Dictionary<ChannelEnum, double>>();
 
         /// <summary>
         /// Максимальная компонента в векторе среднеквадратичного отклонения
         /// </summary>
-        private readonly IList<Tuple<ChannelEnum, float>> _sigmajMax = new List<Tuple<ChannelEnum, float>>();
+        private readonly IList<Tuple<ChannelEnum, double>> _sigmajMax = new List<Tuple<ChannelEnum, double>>();
 
         /// <summary>
         /// Коэффициент при высчитывании gammaj
         /// </summary>
-        private float _coefficient;
+        private double _coefficient;
 
         /// <summary>
         /// Расстояния между всеми парами кластеров
         /// </summary>
-        private readonly IList<Tuple<Cluster, Cluster, float>> _dij = new List<Tuple<Cluster, Cluster, float>>();
+        private readonly IList<Tuple<Cluster, Cluster, double>> _dij = new List<Tuple<Cluster, Cluster, double>>();
 
         #endregion
 
@@ -132,7 +132,7 @@ namespace BusinessLogic.Managers
                 //Шаг 4 алгоритма
                 foreach (var cluster in _z)
                 {
-                    cluster.CenterCluster = new Dictionary<ChannelEnum, float>();
+                    cluster.CenterCluster = new Dictionary<ChannelEnum, double>();
                     foreach (var channel in channels)
                     {
                         cluster.CenterCluster.Add(channel,0f);
@@ -194,11 +194,11 @@ namespace BusinessLogic.Managers
             //Шаг 8
             foreach (var cluster in _z)
             {
-                var dictionary = new Dictionary<ChannelEnum, float>();
+                var dictionary = new Dictionary<ChannelEnum, double>();
                 foreach (var channel in channels)
                 {
-                    var value = cluster.Points.Sum(point => (float) Math.Pow((point.Values[channel] - cluster.CenterCluster[channel]), 2));
-                    value = (float) Math.Sqrt(value/cluster.Points.Count());
+                    var value = cluster.Points.Sum(point => Math.Pow((point.Values[channel] - cluster.CenterCluster[channel]), 2));
+                    value = Math.Sqrt(value/cluster.Points.Count());
                     dictionary.Add(channel,value);
                 }
                 _sigmaj.Add(dictionary);
@@ -207,7 +207,7 @@ namespace BusinessLogic.Managers
             //Шаг 9
             for (var i = 0; i < _z.Count; i++)
             {
-                var max = float.MinValue;
+                var max = double.MinValue;
                 var channel = ChannelEnum.Unknown;
                 foreach (var key in _sigmaj.ElementAt(i).Keys)
                 {
@@ -217,7 +217,7 @@ namespace BusinessLogic.Managers
                         channel = key;
                     }
                 }
-                _sigmajMax.Add(new Tuple<ChannelEnum, float>(channel, max));
+                _sigmajMax.Add(new Tuple<ChannelEnum, double>(channel, max));
             }
 
             //Шаг 10
@@ -256,7 +256,7 @@ namespace BusinessLogic.Managers
             //Шаг 11
             for (var i = 0; i < _z.Count - 1; i++)
             {
-                var tuple = new Tuple<Cluster, Cluster, float>(_z[i], _z[i + 1], EuclideanDistance(_z[i].CenterCluster, _z[i + 1].CenterCluster));
+                var tuple = new Tuple<Cluster, Cluster, double>(_z[i], _z[i + 1], EuclideanDistance(_z[i].CenterCluster, _z[i + 1].CenterCluster));
                 _dij.Add(tuple);
 
                 _z[i].IsJoined = false;
@@ -297,11 +297,11 @@ namespace BusinessLogic.Managers
         /// <param name="pointA">Значения точки А</param>
         /// <param name="pointB">Значения точки B</param>
         /// <returns></returns>
-        private float EuclideanDistance(Dictionary<ChannelEnum, float> pointA, Dictionary<ChannelEnum, float> pointB)
+        private double EuclideanDistance(Dictionary<ChannelEnum, double> pointA, Dictionary<ChannelEnum, double> pointB)
         {
             var keys = pointA.Keys;
-            var result = keys.Sum(key => (float) Math.Pow((pointA[key] - pointB[key]), 2));
-            return (float)Math.Sqrt(result);
+            var result = keys.Sum(key => Math.Pow((pointA[key] - pointB[key]), 2));
+            return Math.Sqrt(result);
         }
 
         /// <summary>
@@ -316,7 +316,7 @@ namespace BusinessLogic.Managers
             var result = new List<Cluster>();
             for (var i = step; i < points.Count(); i += step)
             {
-                var dictinary = new Dictionary<ChannelEnum, float>();
+                var dictinary = new Dictionary<ChannelEnum, double>();
                 for (var j = 0; j < channels.Count(); j++)
                 {
                     var channel = channels.ElementAt(j);
@@ -362,7 +362,7 @@ namespace BusinessLogic.Managers
             }
         }
 
-        private Color GetColorFromNdvi(float ndvi)
+        private Color GetColorFromNdvi(double ndvi)
         {
             Color color = new Color();
